Adding King Piece

+++ Piece/src/Bishop.cpp
@@ -26,8 +26,8 @@ bool Bishop::isValidMove(const Position& initPos, const Position& finalPos) cons
         return false;
     }
 
-    int diffX = abs(initPos.getX() - finalPos.getX());
-    int diffY = abs(initPos.getY() - finalPos.getY());
+    int8_t diffX = abs(initPos.getX() - finalPos.getX());
+    int8_t diffY = abs(initPos.getY() - finalPos.getY());
     if (diffX == diffY)
     {
         return true;
@@ -45,15 +45,15 @@ bool Bishop::isValidMove(const Position& initPos, const Position& finalPos) cons
 std::vector<Position> Bishop::getPath(const Position& initPos, const Position& finalPos) const
 {
     std::vector<Position> path;
-    int pathLength = (abs(initPos.getX() - finalPos.getX()) +
+    int8_t pathLength = (abs(initPos.getX() - finalPos.getX()) +
         abs(initPos.getY() - finalPos.getY())) / 2 + 1;
 
     //Integer.signum(a) provides the sign of a number 1 if positive and -1 if negative.
     //In this case i am considering initPos as the first point and finalPos as second
-    int i_X = Position::sgn(finalPos.getX() - initPos.getX());
-    int i_Y = Position::sgn(finalPos.getY() - initPos.getY());
+    int8_t i_X = Position::sgn(finalPos.getX() - initPos.getX());
+    int8_t i_Y = Position::sgn(finalPos.getY() - initPos.getY());
 
-    for (int cnt = 0; cnt < pathLength; cnt++)
+    for (int8_t cnt = 0; cnt < pathLength; cnt++)
     {
         path.push_back(Position(initPos.getX() + i_X * cnt, initPos.getY() + i_Y * cnt));
     }
